Add a trait for l10n collections
